Add an InviteMembers component
