+++ .eslintrc.js
@@ -15,7 +15,7 @@ module.exports = {
     ecmaVersion: 2020
   },
   rules: {
-    'max-len': ['warn', { code: 110, tabWidth: 4 }],
+    'max-len': ['warn', { code: 120, tabWidth: 4 }],
     'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
     quotes: ['error', 'single', { avoidEscape: true }],
